Disable run-latest-node job for now
We're still using an old version of Storybook that works fine, but uses
Webpack v4, which is no longer supported in latest Node. We should
update, but it's not important from a runtime perspective at all, so we
can ignore it for a while to get the build working first.